fastapi-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
sentence-transformers
httpx
python-dotenv
